+++ team-00/src/main/java/ubc/cosc322/Extras.java
@@ -3,10 +3,10 @@ package ubc.cosc322;
 import java.util.ArrayList;
 
 public class Extras {
-    public static ArrayList<Integer> arrayToArrayList (int[] array) {
+    public static ArrayList<Integer> arrayToArrayList (byte[] array) {
         ArrayList<Integer> returnList = new ArrayList<>(2);
-        returnList.add(0, array[0]);
-        returnList.add(1, array[1]);
+        returnList.add(0, (int)array[0]);
+        returnList.add(1, (int)array[1]);
         return returnList;
     }
 }
